Re #216 - Moved mantid reduction outside of addie.autNOM
